Enhance HexViewer

+++ src/IO/FileAnalyse/SPKFileAnalyse.cpp
@@ -380,7 +380,7 @@ IO::FileAnalyse::FrameDetail *IO::FileAnalyse::SPKFileAnalyse::GetFrameDetail(UO
 	if (pack == 0)
 		return 0;
 
-	NEW_CLASS(frame, IO::FileAnalyse::FrameDetail(pack->fileOfst, pack->packSize));
+	NEW_CLASS(frame, IO::FileAnalyse::FrameDetail(pack->fileOfst, (UInt32)pack->packSize));
 	return frame;
 /*	
 	sb->AppendU64(pack->fileOfst);
